Updating files to achieve automated headless browser testing on Travis-CI
Updating database.py to not ask for input (assume local everytime). Update Selenium test to use chrome without a given path. Update travis.yml to allow things to happen (such as installing the chromedriver and running the database.py file). I expect several things will need to change. Using [this](https://www.ocadotechnology.com/blog/2019/1/14/how-to-use-chromedriver-with-travis-ci-for-selenium-tests) as a reference.